+++ src/spaceObjects/missiles/missileWeapon.cpp
@@ -91,7 +91,7 @@ void MissileWeapon::takeDamage(float damage_amount, DamageInfo info)
         return;
     if (info.type == DT_Heat)
         return;
-    if (random(1,100)<50)
+    if (random(1,100)<80)
         return;
 
     destroy();